requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
ultralytics>=8.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0

